extract steps, improve 1st step
